pyproject.toml
[tool.poetry]
name = "ekyc-core-api"
version = "0.0.1"
description = "eKYC Core API"
authors = ["Khiem Le <theinnocentman1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.0"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
loguru = "^0.7.2"
pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
ultralytics = "^8.2.2"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
